This PR migrates the project to use groups of IPs in their allowlist.
By referring to groups to IP addresses, we can centralize the definition of groups of ip addresses.
If these lists require changing in the future, we can change the definition once and future deploys across all services will automatically include these new IPs.

`1` allowlist(s) have been detected that can be migrated.


**file:helm_deploy/probation-teams/values.yaml**
The effect of applying this PR is as follows:

- The following groups will be applied: `INTERNAL`
- The size of the allowlist defined in this file will change: `8 => 0 (8 removed)`


Merging this PR should not result in any additional IP addresses being added to the allowlist.